+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/PermissionTest.java
@@ -3,7 +3,7 @@
 package com.azure.cosmos;
 
 import com.azure.cosmos.models.PartitionKey;
-import com.azure.cosmos.models.Permission;
+import com.azure.cosmos.implementation.Permission;
 import com.azure.cosmos.models.PermissionMode;
 import org.testng.annotations.Test;
