+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/implementation/ResourceResponseValidator.java
@@ -5,7 +5,6 @@ package com.azure.cosmos.implementation;
 import com.azure.cosmos.models.CompositePath;
 import com.azure.cosmos.models.IndexingMode;
 import com.azure.cosmos.models.ModelBridgeInternal;
-import com.azure.cosmos.models.Permission;
 import com.azure.cosmos.models.PermissionMode;
 import com.azure.cosmos.models.Resource;
 import com.azure.cosmos.models.SpatialSpec;
